[#16] refactor: Home 최상단의 제목과 입력폼을 IntroMain 컴포넌트로 분리
