+++ src/components/intro/IntroMain/index.tsx
@@ -1,0 +1,41 @@
+import React, { FC } from 'react';
+
+import Shell from '@/components/Shell';
+import Jumbotron from '@/components/common/Jumbotron';
+
+const IntroMain: FC = () => {
+  /**
+   * FIXME:
+   * 가상 API 요청을 위해 만든 코드입니다.
+   * 추후 API 연동을 할 때 지워주면 됩니다.
+   */
+  const requestWhenQuestionDone = async () =>
+    new Promise((resolve, reject) => {
+      setTimeout(
+        // () => resolve('Success'),
+        () => reject(new Error('Not valid domain. Telbby is great you know?')),
+        1000,
+      );
+    });
+
+  return (
+    <>
+      <Jumbotron
+        title="Talk"
+        descList={[
+          'Thinking about project feedback?',
+          'Here telbby will listen.',
+        ]}
+      />
+      {/* @TODO 입력폼에 대한 기능 추가가 필요합니다 */}
+      <Shell
+        type="services"
+        requestWhenQuestionDone={requestWhenQuestionDone}
+        width={789}
+        height={208}
+      />
+    </>
+  );
+};
+
+export default IntroMain;
